FabianGamerCoin-($FGC)_MuesliSwap_listing_request
FabianGamerCoin ($FGC) is a funny bond between young gamers and Cardano Native Tokens (both FT & cNFT), that was formerly created for educational purposes, marketing and sponsoring. FGC team aim to explore the benefits of the blockchain and cryptocurrency in the development of profit/non-profit organizations, as well as their use in trading of goods and services. 90% of the supply was sent to the DripDropz team, and the distribution is ongoing (50 $FGC per epoch per elegible wallet, but we are planning to increase/reward holders with an additional 50 $FGC per epoch), and only 7% of the supply will be given to the FGC team. 1% of this amount can be used to sell bundled with Cardano kits NFTs, and the resting supply of the $FGC will be locked for a year, and/or can only be used to provide liquidity to Cardano DEX (we will be grateful to with MuesliSwap to be included). After a year, the members of team will be able to sell or swap from 10 - 40%, (still under discussion) of their FGC. The initial price could be 0.0001 ADA.